# web/src/lib/openapi.snapshot.json
{
  "components": {
    "schemas": {
      "BaseMode": {
        "enum": [
          "survey-in",
          "fixed",
          "off"
        ],
        "title": "BaseMode",
        "type": "string"
      },
      "ConfigBody": {
        "properties": {
          "values": {
            "additionalProperties": true,
            "title": "Values",
            "type": "object"
          }
        },
        "required": [
          "values"
        ],
        "title": "ConfigBody",
        "type": "object"
      },
      "FreezeBody": {
        "properties": {
          "activate": {
            "default": false,
            "title": "Activate",
            "type": "boolean"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "FreezeBody",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModeBody": {
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/BaseMode"
          },
          "site": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site"
          },
          "svin_acc_limit_m": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Svin Acc Limit M"
          },
          "svin_min_duration_s": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Svin Min Duration S"
          }
        },
        "required": [
          "mode"
        ],
        "title": "ModeBody",
        "type": "object"
      },
...
      "SiteBody": {
        "description": "A site given either as ECEF metres or as geodetic degrees plus an ellipsoidal height.",
        "properties": {
          "epoch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Epoch"
          },
          "frame": {
            "default": "ITRF2020",
            "title": "Frame",
            "type": "string"
          },
          "height_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height M"
          },
          "lat": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lon": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lon"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "sigma_m": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sigma M"
          },
          "source": {
            "default": "manual",
            "title": "Source",
            "type": "string"
          },
          "x": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "X"
          },
          "y": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y"
          },
          "z": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Z"
          }
        },
        "required": [
          "name"
        ],
        "title": "SiteBody",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "mtrtk",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/base/mode": {
      "get": {
        "operationId": "get_mode_api_base_mode_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Mode Api Base Mode Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Mode",
        "tags": [
          "base"
        ]
      },
      "put": {
        "description": "Change the base mode, persist it to `.env` and apply it to the receiver.\n\nOne path: `apply_settings_change` validates, refuses `fixed` with no resolvable site (409,\nbefore anything is written), rewrites `.env` and calls `apply_mode()`. A body that changes\nnothing is a no-op - `POST /api/base/survey/restart` is how a running survey is started over.",
        "operationId": "put_mode_api_base_mode_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Put Mode Api Base Mode Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "no base mode manager, or fixed mode with no site to sit on"
          },
          "422": {
            "description": "a value the settings refuse"
          }
        },
        "summary": "Put Mode",
        "tags": [
          "base"
        ]
      }
    },
    "/api/base/sites": {
      "get": {
        "operationId": "list_sites_api_base_sites_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Sites Api Base Sites Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Sites",
        "tags": [
          "base"
        ]
      },
      "post": {
        "description": "Save a site. `{\"site\", \"applied\"}`, the same shape freeze and activate answer with.\n\nA site that has just been added is never the one the receiver is sitting on - adding and\nactivating are separate decisions - so `applied` is false here. It is reported all the same,\nso a client reads one shape from all three routes.",
        "operationId": "add_site_api_base_sites_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Add Site Api Base Sites Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "a site of that name already exists"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Site",
        "tags": [
          "base"
        ]
      }
    },
    "/api/base/sites/{name}": {
      "delete": {
        "operationId": "delete_site_api_base_sites__name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Delete Site Api Base Sites  Name  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "no site of that name"
          },
          "409": {
            "description": "the site is the active one"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Site",
        "tags": [
          "base"
        ]
      }
    },
    "/api/base/sites/{name}/activate": {
      "post": {
        "operationId": "activate_site_api_base_sites__name__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Activate Site Api Base Sites  Name  Activate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "no site of that name"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activate Site",
        "tags": [
          "base"
        ]
      }
    },
...
    "/api/base/survey/freeze": {
      "post": {
        "description": "Save the mean of a completed survey-in as a site, optionally activating it.",
        "operationId": "freeze_survey_api_base_survey_freeze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FreezeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Freeze Survey Api Base Survey Freeze Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "no base mode manager, no valid survey-in, or the name is taken"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Freeze Survey",
        "tags": [
          "base"
        ]
      }
    },
    "/api/base/survey/restart": {
      "post": {
        "description": "Start the survey-in over: TMODE off, then survey-in.\n\nRe-sending the same survey-in parameters does not restart a survey in progress on HPG 1.13,\nso this is not the same request as `PUT /api/base/mode {\"mode\": \"survey-in\"}` - which, when\nnothing changes, deliberately leaves the receiver alone.",
        "operationId": "restart_survey_api_base_survey_restart_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Restart Survey Api Base Survey Restart Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "no base mode manager, not surveying, or the receiver refused"
          }
        },
        "summary": "Restart Survey",
        "tags": [
          "base"
        ]
      }
    },
    "/api/config": {
      "get": {
        "operationId": "get_config_api_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Config Api Config Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Config",
        "tags": [
          "config"
        ]
      },
      "put": {
        "operationId": "put_config_api_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Put Config Api Config Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Config",
        "tags": [
          "config"
        ]
      }
    },
...
  }
}